requirements.txt
aiohttp>=3.9.0
lxml>=5.1.0
motor>=3.3.0
pymongo>=4.6.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
numpy>=1.26.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
playwright>=1.41.0
